requirements.txt
aiogram==3.10.0
SQLAlchemy==2.0.30
asyncpg==0.29.0
python-dotenv==1.0.1
